# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A security analysis tool for JWT tokens. Analyzes JWT structure, claims, and identifies potential security vulnerabilities.",
        "title": "JWT Token Analyzer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/analyze": {
            "post": {
                "description": "Analyzes one or more JWT tokens for security vulnerabilities and best practices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyze"
                ],
                "summary": "Analyze JWT tokens",
                "parameters": [
                    {
                        "description": "JWT tokens to analyze",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis report",
                        "schema": {
                            "$ref": "#/definitions/models.Report"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/decode": {
            "post": {
                "description": "Decodes a JWT token and returns its header, payload, and metadata without security analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decode"
                ],
                "summary": "Decode a JWT token",
                "parameters": [
                    {
                        "description": "JWT token to decode",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DecodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Decoded token",
                        "schema": {
                            "$ref": "#/definitions/handler.DecodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of the service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.AnalyzeRequest": {
            "type": "object",
            "required": [
                "tokens"
            ],
            "properties": {
                "check_expiration": {
                    "type": "boolean",
                    "example": true
                },
                "max_token_lifetime_hours": {
                    "type": "integer",
                    "example": 168
                },
                "tokens": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    ]
                },
                "warn_token_lifetime_hours": {
                    "type": "integer",
                    "example": 24
                }
            }
        },
        "handler.DecodeRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "handler.DecodeResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "header": {
                    "$ref": "#/definitions/models.JWTHeader"
                },
                "payload": {
                    "$ref": "#/definitions/models.DecodedPayload"
                },
                "signature": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "token_info": {
                    "$ref": "#/definitions/models.TokenInfo"
                }
            }
        },
        "models.AnalysisResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "header": {
                    "$ref": "#/definitions/models.JWTHeader"
                },
                "is_valid": {
                    "type": "boolean"
                },
                "payload": {
                    "$ref": "#/definitions/models.DecodedPayload"
                },
                "security_issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SecurityIssue"
                    }
                },
                "signature": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/models.AnalysisSummary"
                },
                "token": {
                    "type": "string"
                },
                "token_info": {
                    "$ref": "#/definitions/models.TokenInfo"
                }
            }
        },
        "models.AnalysisSummary": {
            "type": "object",
            "properties": {
                "critical_count": {
                    "type": "integer"
                },
                "high_count": {
                    "type": "integer"
                },
                "info_count": {
                    "type": "integer"
                },
                "low_count": {
                    "type": "integer"
                },
                "medium_count": {
                    "type": "integer"
                },
                "overall_rating": {
                    "type": "string"
                },
                "security_score": {
                    "type": "string"
                },
                "total_issues": {
                    "type": "integer"
                }
            }
        },
        "models.DecodedPayload": {
            "type": "object",
            "properties": {
                "custom_claims": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "raw_claims": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "standard_claims": {
                    "$ref": "#/definitions/models.StandardClaims"
                }
            }
        },
        "models.JWTHeader": {
            "type": "object",
            "properties": {
                "alg": {
                    "type": "string"
                },
                "cty": {
                    "type": "string"
                },
                "jku": {
                    "type": "string"
                },
                "jwk": {
                    "type": "object"
                },
                "kid": {
                    "type": "string"
                },
                "typ": {
                    "type": "string"
                },
                "x5c": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "x5t": {
                    "type": "string"
                },
                "x5u": {
                    "type": "string"
                }
            }
        },
        "models.Report": {
            "type": "object",
            "properties": {
                "analysis_date": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisResult"
                    }
                }
            }
        },
        "models.SecurityIssue": {
            "type": "object",
            "properties": {
                "affected_field": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "recommendation": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/models.Severity"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Severity": {
            "type": "string",
            "enum": [
                "critical",
                "high",
                "medium",
                "low",
                "info",
                "ok"
            ],
            "x-enum-varnames": [
                "SeverityCritical",
                "SeverityHigh",
                "SeverityMedium",
                "SeverityLow",
                "SeverityInfo",
                "SeverityOK"
            ]
        },
        "models.StandardClaims": {
            "type": "object",
            "properties": {
                "aud": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "iat": {
                    "type": "integer"
                },
                "iss": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "nbf": {
                    "type": "integer"
                },
                "sub": {
                    "type": "string"
                }
            }
        },
        "models.TokenInfo": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "has_signature": {
                    "type": "boolean"
                },
                "is_expired": {
                    "type": "boolean"
                },
                "is_not_yet_valid": {
                    "type": "boolean"
                },
                "issued_at": {
                    "type": "string"
                },
                "not_before": {
                    "type": "string"
                },
                "parts_count": {
                    "type": "integer"
                },
                "signature_present": {
                    "type": "boolean"
                },
                "time_since_expiry": {
                    "type": "string"
                },
                "time_until_expiry": {
                    "type": "string"
                },
                "token_lifetime": {
                    "type": "string"
                }
            }
        }
    }
}